requirements.txt
wd14-tagger-api
onnxruntime-gpu
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.3
databases==0.8.0
decorator==5.1.1
deepdanbooru==1.0.2
Deprecated==1.2.14
exceptiongroup==1.2.0
fabric==3.2.2
fastapi==0.109.2
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
humanize==4.9.0
idna==3.6
imageio==2.34.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.16.0
onnxruntime-gpu==1.17.0
opencv-python==4.8.0.76
packaging==23.2
pandas==2.1.1
paramiko==3.4.0
Pillow==10.0.1
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyNaCl==1.5.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
Redis-Sentinel-Url==1.0.1
requests==2.31.0
rq==1.15.1
rq-dashboard==0.6.7.2
rq-win==0.4.0
rqmonitor==1.0.6
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.36.3
sympy==1.12
tifffile==2024.2.12
times==0.7
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.66.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
wd14-tagger-api==0.0.1
Werkzeug==3.0.1
wrapt==1.16.0